refactor integration type info
