+++ clients/admin-ui/src/features/integrations/ConnectionCategory.ts
@@ -1,0 +1,4 @@
+export enum ConnectionCategory {
+  DATA_WAREHOUSE = "Data Warehouse",
+  DATABASE = "Database",
+}
